refactor: ChildrenDropdownFilter에  option type과 service 분리 적용
